+++ admin/download_installers.py
@@ -99,6 +99,10 @@ PATTERNS = {
     (EE_MASTER, ),
     'tests/test_dcos_e2e/backends/docker/test_distributions.py::TestCoreOS::test_oss':  # noqa: E501
     (OSS_MASTER, ),
+    'tests/test_dcos_e2e/backends/docker/test_distributions.py::TestFlatcar::test_enterprise':  # noqa: E501
+    (EE_MASTER, ),
+    'tests/test_dcos_e2e/backends/docker/test_distributions.py::TestFlatcar::test_oss':  # noqa: E501
+    (OSS_MASTER, ),
     'tests/test_dcos_e2e/backends/docker/test_distributions.py::TestUbuntu1604::test_oss':  # noqa: E501
     (OSS_MASTER, ),
     'tests/test_dcos_e2e/backends/docker/test_distributions.py::TestUbuntu1604::test_enterprise':  # noqa: E501